EES-3281 Extract `ChartReferenceLinesConfiguration` from chart axis config
